+++ CHANGELOG.md
@@ -4,8 +4,9 @@ This project follows [semantic versioning](http://semver.org).
 
 ### UNRELEASED
 
-- [changed] Changed Rust version to 1.64.0
+- [changed] Changed Rust version to 1.73.0
 - [changed] Updated dependencies
+- [fixed] Fix error when IPv6 is not available
 
 ### v2.3.0 (2021-12-23)